FIX: Modify selected answer retrieval method
The selected answer was determined by retrieving the innerHTML of the
answer button which was selected. Symbols (e.g. '&') would be converted
to a character entity references (e.g. &amp;) which would cause the
answer comparison to fail.

Modified the code so the checkAnswer function
so the innerText of the element is retrieved for comparison.