FIX: Remove question count check
Originally introduced in commit
163fab441c4bb201d61ed44dc8f2c29e88134fc2, further testing and mentor
feedback revealed that this check was not returning actionable
information. While the function would retrieve the number of questions
available for a specific category, this would include questions of any
type whereas this application only uses multiple choice questions.

Future development would aim to incorporate all question types in to the
quiz so this information would be useable.

The application now relies on the API response code obtained during data
retrieval. If a response code of 1 is returned not enough questions were
available in the database and this throws an error which is reported to
the user.